Updating to 0.4.0 Joint. Project reorganization.
